removed ClusterWithLeds - too specific
